move env files to tasks
