+++ .gitignore
@@ -28,6 +28,9 @@ hydra_outputs/
 
 # Usual python .gitignore
 
+# Mac OS X files
+.DS_Store
+
 # Byte-compiled / optimized / DLL files
 __pycache__/
 *.py[cod]
