--- docs/api_documentation/environments.md
@@ -1,3 +1,0 @@
-# Environments
-
-::: qdax.environments
