+++ qdax/core/emitters/qdpg_emitter.py
@@ -20,7 +20,7 @@ from qdax.core.emitters.qpg_emitter import QualityPGConfig, QualityPGEmitter
 from qdax.core.emitters.repertoire_selectors.selector import Selector
 from qdax.core.emitters.standard_emitters import MixingEmitter
 from qdax.custom_types import Reward, StateDescriptor
-from qdax.environments.base_wrappers import QDEnv
+from qdax.tasks.brax.v1.base_wrappers import QDEnv
 
 
 @dataclass
