+++ qdax/utils/train_seq2seq.py
@@ -17,7 +17,7 @@ from flax.training import train_state
 from qdax.core.containers.unstructured_repertoire import UnstructuredRepertoire
 from qdax.core.neuroevolution.networks.seq2seq_networks import Seq2seq
 from qdax.custom_types import Params, RNGKey
-from qdax.environments.descriptor_extractors import AuroraExtraInfoNormalization
+from qdax.environments import AuroraExtraInfoNormalization
 
 
 def get_model(
