+++ tests/baselines_test/dcrlme_test.py
@@ -14,8 +14,8 @@ from qdax.core.neuroevolution.buffers.buffer import DCRLTransition
 from qdax.core.neuroevolution.networks.networks import MLP, MLPDC
 from qdax.custom_types import EnvState, Params, RNGKey
 from qdax.environments import descriptor_extractor
-from qdax.environments.wrappers import ClipRewardWrapper, OffsetRewardWrapper
-from qdax.tasks.brax_envs import scoring_function_brax_envs
+from qdax.tasks.brax.v1.env_creators import scoring_function_brax_envs
+from qdax.tasks.brax.v1.wrappers import ClipRewardWrapper, OffsetRewardWrapper
 from qdax.utils.metrics import default_qd_metrics
